+++ simpleble/src/backends/dongl/usb/UsbHelperWindows.cpp
@@ -1,0 +1,13 @@
+#include "UsbHelper.h"
+
+namespace SimpleBLE {
+namespace Dongl {
+namespace USB {
+
+std::vector<std::string> UsbHelper::get_dongl_devices() {
+    return {};
+}
+
+}  // namespace USB
+}  // namespace Dongl
+}  // namespace SimpleBLE